+++ algo_handler_ops_list.c
@@ -1,3 +1,5 @@
+#include <stdlib.h>
+
 extern struct beast_ops des_handler_ops;
 
 struct beast_ops *ops_handler_list[] = {
